+++ api/crates/perception_infra/src/lib.rs
@@ -8,6 +8,7 @@ mod transient_dataset_version_repository;
 mod transient_sample_repository;
 mod transient_training_job_queue;
 mod transient_training_job_repository;
+mod transient_training_metric_repository;
 
 pub use local_sample_storage::LocalSampleStorage;
 pub use transient_annotation_repository::TransientAnnotationRepository;
@@ -16,5 +17,6 @@ pub use transient_dataset_version_repository::TransientDatasetVersionRepository;
 pub use transient_sample_repository::TransientSampleRepository;
 pub use transient_training_job_queue::TransientTrainingJobQueue;
 pub use transient_training_job_repository::TransientTrainingJobRepository;
+pub use transient_training_metric_repository::TransientTrainingMetricRepository;
 
 pub const CRATE_NAME: &str = "perception_infra";